// src/app/demos/demo-test/tests.types.ts
export type TestCase = Readonly<{
  suite: string;
  name: string;
  meta?: Record<string, string>;
  run: () => void | RunCaseRet | Promise<void | RunCaseRet>;
}>;

export type RunCaseRet = Readonly<{
  metaPatch?: Record<string, string>;
  assertRows?: readonly TestAssertRow[];
}>;

export type TestSuite = Readonly<{
  suite: string;
  cases: readonly TestCase[];
}>;